refactor: create a new and more organized structure on the repository
