+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "files.exclude": {
+    "**/dist": true
+  },
+  "files.insertFinalNewline": true,
+  "files.trimTrailingWhitespace": true
+}
