+++ examples/index.html
@@ -76,6 +76,9 @@
         <a href="street-view-pano.html">Street View Panorama</a> (Experimental).
       </li>
       <li>
+        <a href="street-view-pano-markers.html">Street View Panorama with Markers</a> (Experimental).
+      </li>
+      <li>
         <a href="resize-bus.html">Resize Bus</a>.
       </li>
       <li>
