+++ src/components/cluster.js
@@ -8,7 +8,7 @@
   https://github.com/googlemaps/v3-utility-library/blob/master/markerclustererplus/src/markerclusterer.js
 **/
 import MarkerClusterer from 'marker-clusterer-plus'
-import mapElementFactory from './mapElementFactory.js'
+import mapElementFactory from '../factories/map-element'
 
 const props = {
   maxZoom: {
