+++ src/utils/lazy-value.js
@@ -1,7 +1,7 @@
 // This piece of code was orignally written by sindresorhus and can be seen here
 // https://github.com/sindresorhus/lazy-value/blob/master/index.js
 
-export default fn => {
+export default (fn) => {
   let called = false
   let ret
