+++ src/utils/simulate-arrow-down.js
